### 2025-11-10 #8986 @davidmeunier79
@rkurchin Many thanks for your advice; I have done the required modifications; One DOI was missing; it appears the previous two invalid DOIs are not properly formatted, hence required the use of "url" field in the bibtex, now they are in the "skip DOIs" rather than "invalid DOIs".

 Please let me know if that version now fits the expected requirements.
