### 2025-12-15 49d9554
Add padded wmgm (#316)
* stereo_padded_wmgm_mask
