### 2025-12-15 b1990c2
added nonlin_warp_file and nonlin_invwarp_file as non mandatory outputs
