### 2025-12-17 99e76ad
modif params lower3
