+++ macapype/pipelines/full_pipelines.py
@@ -251,7 +251,7 @@ def create_brain_segment_from_mask_pipe(
     # creating outputnode
     outputnode = pe.Node(
         niu.IdentityInterface(
-            fields=["segmented_file",
+            fields=["segmented_file", "stereo_parcel", "stereo_parcel_gm",
                     "threshold_gm", "threshold_wm", "threshold_csf",
                     "prob_gm", "prob_wm", "prob_csf"]),
         name='outputnode')
@@ -401,25 +401,6 @@ def create_brain_segment_from_mask_pipe(
                     reg, 'nonlin_invwarp_file',
                     register_parcel_to_nat, "field_file")
 
-#
-#
-#
-#                 # seg
-#                 register_parcel_to_nat = pe.Node(
-#                     fsl.ApplyXFM(), name="register_parcel_to_nat")
-#                 register_parcel_to_nat.inputs.interp = "nearestneighbour"
-#
-#                 register_parcel_to_nat.inputs.in_file = params_template[
-#                     "template_parcel"]
-#                 brain_segment_pipe.connect(
-#                     inputnode, 'masked_debiased_T1',
-#                     register_parcel_to_nat, 'reference')
-#
-#                 brain_segment_pipe.connect(
-#                     reg, 'inv_transfo_file',
-#                     register_parcel_to_nat, "in_matrix_file")
-#
-
         elif "reg_f3d" in params:
             # Iterative registration to the template
             reg_f3d = pe.Node(
@@ -437,36 +418,7 @@ def create_brain_segment_from_mask_pipe(
             brain_segment_pipe.connect(
                 inputnode, 'masked_debiased_T1',
                 reg_f3d, 'fmask_file')
-#
-#             # merge_tuple
-#             merge_tuple =  pe.Node(
-#                 Merge(3), name="merge_tuple")
-#
-#             brain_segment_pipe.connect(
-#                 reg_f3d, 'masked_debiased_T1',
-#                 merge_tuple, 'in1')
-#
-#             brain_segment_pipe.connect(
-#                 reg, 'nonlin_invwarp_file',
-#                 merge_tuple, "in2")
-#
-#             # inv_transfo
-#             inv_transfo = pe.Node(
-#                 RegTransform(), name="inv_transfo")
-#
-#             inv_transfo.inputs.interp = "nn"
-#
-#             inv_transfo.inputs.in_file = params_template[
-#                 "template_seg"]
-#
-#             inv_transfo.connect(
-#                 merge_tuple, 'out',
-#                 register_seg_to_nat, 'inv_nrr_input')
-#
-#             brain_segment_pipe.connect(
-#                 reg, 'nonlin_invwarp_file',
-#                 register_seg_to_nat, "field_file")
-#
+
             if "template_seg" in params_template.keys():
 
                 # seg
@@ -538,7 +490,6 @@ def create_brain_segment_from_mask_pipe(
                     reg, 'inv_transfo_file',
                     register_csf_to_nat, "in_matrix_file")
 
-
             if "template_parcel" in params_template.keys():
 
                 register_parcel_to_nat = pe.Node(
@@ -560,6 +511,9 @@ def create_brain_segment_from_mask_pipe(
         else:
             print("##### Error, no coregistration method is defined")
             return brain_segment_pipe
+
+    brain_segment_pipe.connect(register_parcel_to_nat, 'out_file',
+                                   outputnode, 'stereo_parcel')
 
     # ants Atropos
     if "template_seg" in params_template.keys():
@@ -621,6 +575,24 @@ def create_brain_segment_from_mask_pipe(
         inputnode, 'masked_debiased_T1',
[…]
