### 2025-12-18 be94b11
merge dseg dtype
