### 2025-12-18 a19d6ac
np.int
