### 2025-12-18 38c2e2f
np.int32
