### 2025-12-18 1793066
threshold gm
