+++ macapype/pipelines/rename.py
@@ -427,22 +427,22 @@ def rename_all_brain_derivatives(params, main_workflow, segment_pnh_pipe,
 
         if "parcel_gm" in params["brain_segment_pipe"]:
 
-        # rename parcel_gm
-        rename_stereo_parcel_gm = pe.Node(
-            niu.Rename(),
-            name="rename_stereo_parcel_gm")
-        rename_stereo_parcel_gm.inputs.format_string = \
-            pref_deriv + "_space-stereo_desc-parcelgm_dseg"
-        rename_stereo_parcel_gm.inputs.parse_string = parse_str
-        rename_stereo_parcel_gm.inputs.keep_ext = True
+            # rename parcel_gm
+            rename_stereo_parcel_gm = pe.Node(
+                niu.Rename(),
+                name="rename_stereo_parcel_gm")
+            rename_stereo_parcel_gm.inputs.format_string = \
+                pref_deriv + "_space-stereo_desc-parcelgm_dseg"
+            rename_stereo_parcel_gm.inputs.parse_string = parse_str
+            rename_stereo_parcel_gm.inputs.keep_ext = True
 
-        main_workflow.connect(
-            segment_pnh_pipe, 'outputnode.stereo_parcel_gm',
-            rename_stereo_parcel_gm, 'in_file')
+            main_workflow.connect(
+                segment_pnh_pipe, 'outputnode.stereo_parcel_gm',
+                rename_stereo_parcel_gm, 'in_file')
 
-        main_workflow.connect(
-            rename_stereo_parcel_gm, 'out_file',
-            datasink, '@stereo_parcel_gm')
+            main_workflow.connect(
+                rename_stereo_parcel_gm, 'out_file',
+                datasink, '@stereo_parcel_gm')
 
 
         if "pad_template" in params["short_preparation_pipe"].keys():
