+++ macapype/pipelines/full_pipelines.py
@@ -498,10 +498,6 @@ def create_brain_segment_from_mask_pipe(
             print("##### Error, no coregistration method is defined")
             return brain_segment_pipe
 
-
-    #outputnode
-    brain_segment_pipe.connect(register_parcel_to_nat, 'out_file',
-                               outputnode, 'parcel')
     # ants Atropos
     if "template_seg" in params_template.keys():
 
@@ -779,7 +775,6 @@ def create_full_T1T2_subpipes(
                     "stereo_gen_5tt",
                     "native_gen_5tt",
 
-                    "stereo_parcel",
                     "stereo_parcel_gm",
                     "stereo_padded_parcel_gm"
                     "native_parcel_gm",
@@ -1324,9 +1319,6 @@ def create_full_T1T2_subpipes(
 
     seg_pipe.connect(brain_segment_pipe, 'outputnode.parcel_gm',
                      outputnode, 'stereo_parcel_gm')
-
-    seg_pipe.connect(brain_segment_pipe, 'outputnode.parcel',
-                     outputnode, 'stereo_parcel')
 
 
     if pad and space == "native":
