### 2025-12-22 b012dc7
native parcel_gm
