+++ macapype/pipelines/rename.py
@@ -876,6 +876,26 @@ def rename_all_brain_derivatives(params, main_workflow, segment_pnh_pipe,
                 rename_native_segmented_brain_mask, 'out_file',
                 datasink, '@native_segmented_brain_mask')
 
+            if "parcel_gm" in params["brain_segment_pipe"]:
+
+                rename_native_padded_parcel_gm = pe.Node(
+                    niu.Rename(),
+                    name="rename_native_padded_parcel_gm")
+                rename_native_padded_parcel_gm.inputs.format_string = \
+                    pref_deriv + "_space-native_desc-pad_desc-parcelgm_dseg"
+                rename_native_padded_parcel_gm.inputs.parse_string = \
+                    parse_str
+                rename_native_padded_parcel_gm.inputs.keep_ext = True
+
+                main_workflow.connect(
+                    segment_pnh_pipe,
+                    'outputnode.native_padded_parcel_gm',
+                    rename_native_padded_parcel_gm, 'in_file')
+
+                main_workflow.connect(
+                    rename_native_padded_parcel_gm, 'out_file',
+                    datasink, '@native_padded_parcel_gm')
+
             # rename 5tt
             if "export_5tt_pipe" in params.keys():
                 rename_native_gen_5tt = pe.Node(
