### 2025-12-22 6fa7b46
no native_parcel_gm
