### 2025-12-22 c5e43a8
intermediate params
