### 2025-12-22 2e6302f
using linear flirt and checking
