+++ macapype/nodes/segment.py
@@ -248,7 +248,7 @@ def merge_imgs(list_img_files):
             path, fname, ext = split_f(img_file)
             img = nib.load(img_file)
             img_data = img.get_fdata()
-            new_img_data = np.zeros(shape=img_data.shape, dtype = np.int32)
+            new_img_data = np.zeros(shape=img_data.shape, dtype=np.int32)
             new_img_data[img_data != 0] = 1
 
         else:
