+++ macapype/pipelines/full_pipelines.py
@@ -10,7 +10,7 @@ from nipype.interfaces import fsl
 from nipype.interfaces.ants.utils import ImageMath
 
 from nipype.interfaces.niftyreg.reg import RegF3D
-from nipype.interfaces.niftyreg.regutils import RegResample, RegTransform
+from nipype.interfaces.niftyreg.regutils import RegResample
 
 from ..utils.utils_nodes import NodeParams
 
@@ -31,7 +31,7 @@ from .segment import (create_old_segment_pipe,
 
 from .correct_bias import create_masked_correct_bias_pipe
 
-from .register import (create_register_NMT_pipe, create_reg_seg_pipe)
+from .register import (create_register_NMT_pipe)
 
 from .extract_brain import create_extract_pipe
 
@@ -282,9 +282,8 @@ def create_brain_segment_from_mask_pipe(
 
             reg.inputs.refb_file = params_template["template_brain"]
 
-#
-#             reg.inputs.refw_file = params_template["template_head"]
-#             reg.inputs.k = True
+            # reg.inputs.refw_file = params_template["template_head"]
+            # reg.inputs.k = True
 
             brain_segment_pipe.connect(
                 inputnode, 'debiased_T1',
@@ -368,31 +367,8 @@ def create_brain_segment_from_mask_pipe(
                     reg, 'inv_transfo_file',
                     register_csf_to_nat, "in_matrix_file")
 
-
             if "template_parcel" in params_template.keys() \
-                and "parcel_gm" in params:
-                #
-                # # nonlin flirt
-                # # seg
-                # register_parcel_to_nat = pe.Node(
-                #     fsl.ApplyWarp(), name="register_parcel_to_nat")
-                #
-                # register_parcel_to_nat.inputs.interp = "nn"
-                #
-                # register_parcel_to_nat.inputs.in_file = params_template[
-                #     "template_parcel"]
-                # brain_segment_pipe.connect(
-                #     inputnode, 'masked_debiased_T1',
-                #     register_parcel_to_nat, 'ref_file')
-                #
-                # brain_segment_pipe.connect(
-                #     reg, 'nonlin_invwarp_file',
-                #     register_parcel_to_nat, "field_file")
-                #
-                # #
-                # # brain_segment_pipe.connect(
-                # #     reg, 'inv_transfo_file',
-                # #     register_parcel_to_nat, "premat")
+                    and "parcel_gm" in params:
 
                 # flirt version
                 register_parcel_to_nat = pe.Node(
@@ -410,16 +386,15 @@ def create_brain_segment_from_mask_pipe(
                     reg, 'inv_transfo_file',
                     register_parcel_to_nat, "in_matrix_file")
 
-
         elif "reg_f3d" in params:
             # Iterative registration to the template
             reg_f3d = pe.Node(
                 RegF3D(),
                 name='reg_f3d')
 
-            reg_f3d.inputs.rmask_file= params_template["template_brain"]
+            reg_f3d.inputs.rmask_file = params_template["template_brain"]
 
-            reg_f3d.inputs.ref_file  = params_template["template_head"]
+            reg_f3d.inputs.ref_file = params_template["template_head"]
 
             brain_segment_pipe.connect(
                 inputnode, 'debiased_T1',
@@ -448,62 +423,13 @@ def create_brain_segment_from_mask_pipe(
                     reg_f3d, 'invcpp_file',
                     register_seg_to_nat, "trans_file")
             else:
-
-                #TODO
+                # TODO
+                print("**** register split template files \
+                    is not impremented for template reg_f3d ****")
                 pass
-                # # gm
[…]
