+++ macapype/pipelines/rename.py
@@ -360,6 +360,13 @@ def rename_all_brain_derivatives(params, main_workflow, segment_pnh_pipe,
 
     if "brain_segment_pipe" in params or "brain_old_segment_pipe" in params:
 
+        if "brain_segment_pipe" in params.keys():
+            seg_pipe_name = "brain_segment_pipe"
+
+        elif "brain_old_segment_pipe" in params.keys():
+            seg_pipe_name = "brain_old_segment_pipe"
+
+
         # rename prob_wm
         rename_stereo_prob_wm = pe.Node(niu.Rename(),
                                         name="rename_stereo_prob_wm")
@@ -425,7 +432,7 @@ def rename_all_brain_derivatives(params, main_workflow, segment_pnh_pipe,
             rename_stereo_segmented_brain_mask, 'out_file',
             datasink, '@stereo_segmented_brain_mask')
 
-        if "parcel_gm" in params["brain_segment_pipe"]:
+        if "parcel_gm" in params[seg_pipe_name]:
 
             # rename parcel_gm
             rename_stereo_parcel_gm = pe.Node(
@@ -464,7 +471,7 @@ def rename_all_brain_derivatives(params, main_workflow, segment_pnh_pipe,
                 rename_stereo_padded_segmented_brain_mask, 'out_file',
                 datasink, '@stereo_padded_segmented_brain_mask')
 
-            if "parcel_gm" in params["brain_segment_pipe"]:
+            if "parcel_gm" in params[seg_pipe_name]:
 
                 rename_stereo_padded_parcel_gm = pe.Node(
                     niu.Rename(),
