+++ macapype/pipelines/full_pipelines.py
@@ -9,6 +9,7 @@ from nipype.interfaces import fsl
 
 from nipype.interfaces.ants.utils import ImageMath
 
+from nipype.interfaces.niftyreg.reg import RegF3D
 from nipype.interfaces.niftyreg.regutils import RegResample
 
 from ..utils.utils_nodes import NodeParams
@@ -30,7 +31,7 @@ from .segment import (create_old_segment_pipe,
 
 from .correct_bias import create_masked_correct_bias_pipe
 
-from .register import (create_register_NMT_pipe, create_reg_seg_pipe)
+from .register import (create_register_NMT_pipe)
 
 from .extract_brain import create_extract_pipe
 
@@ -250,7 +251,7 @@ def create_brain_segment_from_mask_pipe(
     # creating outputnode
     outputnode = pe.Node(
         niu.IdentityInterface(
-            fields=["segmented_file",
+            fields=["segmented_file", "parcel", "parcel_gm",
                     "threshold_gm", "threshold_wm", "threshold_csf",
                     "prob_gm", "prob_wm", "prob_csf"]),
         name='outputnode')
@@ -281,9 +282,8 @@ def create_brain_segment_from_mask_pipe(
 
             reg.inputs.refb_file = params_template["template_brain"]
 
-            if "nonlin_reg" in params["reg"]:
-                reg.inputs.refw_file = params_template["template_head"]
-                reg.inputs.k = True
+            # reg.inputs.refw_file = params_template["template_head"]
+            # reg.inputs.k = True
 
             brain_segment_pipe.connect(
                 inputnode, 'debiased_T1',
@@ -301,18 +301,19 @@ def create_brain_segment_from_mask_pipe(
 
                 # seg
                 register_seg_to_nat = pe.Node(
-                    fsl.ApplyXFM(), name="register_seg_to_nat")
-                register_seg_to_nat.inputs.interp = "nearestneighbour"
+                    fsl.ApplyWarp(), name="register_seg_to_nat")
+
+                register_seg_to_nat.inputs.interp = "nn"
 
                 register_seg_to_nat.inputs.in_file = params_template[
                     "template_seg"]
                 brain_segment_pipe.connect(
                     inputnode, 'masked_debiased_T1',
-                    register_seg_to_nat, 'reference')
+                    register_seg_to_nat, 'ref_file')
 
                 brain_segment_pipe.connect(
-                    reg, 'inv_transfo_file',
-                    register_seg_to_nat, "in_matrix_file")
+                    reg, 'nonlin_invwarp_file',
+                    register_seg_to_nat, "field_file")
 
             else:
                 # gm
@@ -365,6 +366,87 @@ def create_brain_segment_from_mask_pipe(
                 brain_segment_pipe.connect(
                     reg, 'inv_transfo_file',
                     register_csf_to_nat, "in_matrix_file")
+
+            if "template_parcel" in params_template.keys() \
+                    and "parcel_gm" in params:
+
+                # flirt version
+                register_parcel_to_nat = pe.Node(
+                    fsl.ApplyXFM(), name="register_parcel_to_nat")
+
+                register_parcel_to_nat.inputs.interp = "nearestneighbour"
+
+                register_parcel_to_nat.inputs.in_file = params_template[
+                    "template_parcel"]
+                brain_segment_pipe.connect(
+                    inputnode, 'masked_debiased_T1',
+                    register_parcel_to_nat, 'reference')
+
+                brain_segment_pipe.connect(
+                    reg, 'inv_transfo_file',
+                    register_parcel_to_nat, "in_matrix_file")
+
+        elif "reg_f3d" in params:
+            # Iterative registration to the template
+            reg_f3d = pe.Node(
+                RegF3D(),
+                name='reg_f3d')
+
[…]
