### 2026-01-27 c360478
Modification to take into account new guidelines:Software Design` , `Research Impact Statement` section in paper and `AI usage disclosure` sections
