### 2026-01-28 061763b
Minor spelling and grammatical fixes to JOSS submission
