### 2026-01-29 #8986 @katjaq
Hi @adamltyson . Thank you for the invitation! I am excited to review this submission for JOSS!
