### 2026-02-19 #8986 @adamltyson
Hi @emdupre, @neurabenn, @wmvanvliet, just checking in again to see if you would be able to review this submission for JOSS?
