### 2026-03-02 #8986 @adamltyson
Hi @Mmasoud1, @elliohow, @paquiteau, @donishadsmith, @htwangtw, @jsheuni 👋 Would you be able to review this submission to JOSS?

If you haven't reviewed for JOSS before, we carry out our checklist-driven reviews here in GitHub issues following [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). If you have any questions, let me know.
