### 2026-03-17 #10203 @editorialbot
[REVIEW]: Macapype: An open multi-software framework for non-human primate brain anatomical MRI processing
**Submitting author:** @davidmeunier79 (<a href="http://orcid.org/0000-0002-5812-6138">David Meunier</a>)
**Repository:** https://github.com/Macatools/macapype
**Branch with paper.md** (empty if default branch): paper_JOSS
**Version:** v0.7.1
