### 2026-03-18 #10203 @paquiteau
> **License info**:
>
> ✅ License found: `MIT License` (Valid open source [OSI approved](https://opensource.org/licenses) license)

Nitpick: This is wrong, the license used in the repository is BSD-3 (which is also OSI admissible)
