### 2026-03-19 #10203 @adamltyson
The main branch is BSD-3, but the JOSS submission branch is MIT. Both are fine, but @davidmeunier79 it may simplify the review if you could make sure the only difference between the main and JOSS branches are the paper itself.
