### 2026-03-23 bf8cfa5
licence MIT for main as well (#319)
