### 2026-03-23 #10203 @davidmeunier79
@adamltyson @paquiteau Thanks for noticing mistake, this is now done (both version in branches master and paper_JOSS have MIT licence)
