### 2026-08-03 4bbb9ce
correction of the docker pull command
