### 2026-08-03 7ea0d19
first version of CONTRIBUTING.md
