### 2026-08-03 02b097a
add link to CONTRIBUTING
