### 2026-08-03 d179664
added badges
