### 2026-08-03 491667f
modified link in prospect to the link in macatools
