### 2026-08-03 e983096
Corrections joss (#321)
* licence MIT for main as well

* correction of the docker pull command

* first version of CONTRIBUTING.md

* add link to CONTRIBUTING

* added badges

* removed TST

* modified link in prospect to the link in macatools

* minor correction : link to macatools not davidmeunier79 github
