+++ README.md
@@ -16,14 +16,14 @@ python setup.py develop
 
 # Docker image
 ```shell
-docker pull macatools/macapype/latest
+docker pull macatools/macapype:latest
 ```
 
 # Documentation
 
 https://macatools.github.io/macapype/
 
-# Related to project:
+# External documentation related to project:
 
 [A guide for PNH MRI processing](https://github.com/PRIME-RE/prime-re.github.io/wiki), and in particular, the session on [structural MRI processing](https://github.com/PRIME-RE/prime-re.github.io/wiki/Structural-processing)
 
@@ -31,6 +31,8 @@ https://macatools.github.io/macapype/
 
 
 # Forum macapype_users
-
 The macapype_users can be be found [here](https://framateam.org/signup_user_complete/?id=ebtxf45nmiyqmkm8e6oh9qjsoe) (requires a framateam/framagit account, but should be accessible with a github or bitbucket account)
 
+
+# Contribution to project
+Willing to contribute to the project: See the instruction in [CONTRUBUTING.md] (https://github.com/macatools/macapype/blob/master/CONTRUBUTING.md)
