+++ _sources/auto_examples/index.rst.txt
@@ -11,6 +11,10 @@ Examples Gallery
 
 
 
+.. raw:: html
+
+  <div id='sg-tag-list' class='sphx-glr-tag-list'></div>
+
 
 .. raw:: html
 
