### 2026-08-03 ca6f966
Fix contributing (#322)
* licence MIT for main as well

* added link to contributing in doc

* bug link to CONTRIBUTING.md
