### 2026-08-03 8828dcb
delete pytest step from the doc
