### 2026-08-11 241f5cb
Useful tips (#325)
* added utils.rst

* Reorient

* other possibilities at the beginning of the page
