+++ macapype/utils/tests/test_utils_nodes.py
@@ -10,32 +10,57 @@ from macapype.utils.utils_nodes import NodeParams, MapNodeParams, ParseParams
 from nipype.interfaces.base import traits
 import nipype.interfaces.fsl as fsl
 
-data_path = load_test_data("data_test_macaque")
-
-T1_file = op.join(data_path, "sub-Apache_ses-01_T1w.nii")
-T2_file = op.join(data_path, "sub-Apache_ses-01_T2w.nii")
-
 
 def test_NodeParams_init():
+    try:
+
+        data_path = load_test_data("data_test_macaque")
+
+    except AssertionError as e:
+        pytest.skip(f"Test data not available: {e}")
+
+    T1_file = op.join(data_path, "sub-Apache_ses-01_T1w.nii")
+    T2_file = op.join(data_path, "sub-Apache_ses-01_T2w.nii")
+
     params = {"t_file": T1_file, "t2_file": T2_file, "aT2": True}
     with pytest.raises(AssertionError):
-        bet_crop = NodeParams(interface=T1xT2BET(), params=params,  # noqa
-                              name="bet_crop")
+        bet_crop = NodeParams(interface=T1xT2BET(),
+                                params=params,
+                                name="bet_crop")
+        bet_crop.run()
 
 
 def test_NodeParams_load_inputs_from_dict():
+    try:
+        data_path = load_test_data("data_test_macaque")
+
+    except AssertionError as e:
+        pytest.skip(f"Test data not available: {e}")
+
+    T1_file = op.join(data_path, "sub-Apache_ses-01_T1w.nii")
+    T2_file = op.join(data_path, "sub-Apache_ses-01_T2w.nii")
+
     params = {"t_file": T1_file, "t2_file": T2_file, "aT2": True}
     bet_crop = NodeParams(interface=T1xT2BET(), name="bet_crop")
+
     with pytest.raises(AssertionError):
         bet_crop.load_inputs_from_dict(params)
 
-
 def test_MapNodeParams():
+
     params = {"crop": {"args": "88 144 14 180 27 103"}}
 
     crop_bb = MapNodeParams(fsl.ExtractROI(), name='crop_bb',
                             params=parse_key(params, "crop"),
                             iterfield=["in_file"])
+    try:
+        data_path = load_test_data("data_test_macaque")
+
+    except AssertionError as e:
+        pytest.skip(f"Test data not available: {e}")
+
+    T1_file = op.join(data_path, "sub-Apache_ses-01_T1w.nii")
+    T2_file = op.join(data_path, "sub-Apache_ses-01_T2w.nii")
 
     crop_bb.inputs.in_file = [T1_file, T2_file]
 
