+++ docs/utils.rst
@@ -1,0 +1,179 @@
+:orphan:
+
+.. _utils:
+
+Useful tips (before processing)
+_______________________________
+
+
+If default pipelines are failing, it may be due to original images being non-optimal for processing. Please check axis labels, orientation , and cropping before retrying to run macapype. Here we will focus on a method using FSL and FSLeyes. Other possibilities to perform interactively labels modification and images orientation include tools such as  `Reorient <https://neuroanatomy.github.io/reorient>`_ .
+
+**************************************
+How to orient the axis labels your images ?
+**************************************
+
+In particular, the image should be in neurological convention, i.e. left label on the right of the image in FSL coronal representation, and x values increasion from left to right.
+
+.. image:: ./img/crop/wrong_orient_mark.png
+    :scale: 50%
+    :align: center
+
+If is not the case, it is possible to use FSL use the following procedure to swap the X-axis
+
+.. code-block:: bash
+
+    cp Orig/sub-Iroquoise_ses-01_T1w.nii.gz .
+    fslreorient2std sub-Iroquoise_ses-01_T1w.nii.gz sub-Iroquoise_ses-01_T1w_reorient.nii.gz
+    fslorient -forceneurological sub-Iroquoise_ses-01_T1w_reorient.nii.gz
+    fslswapdim sub-Iroquoise_ses-01_T1w_reorient.nii.gz -x y z sub-Iroquoise_ses-01_T1w_reorient_swapped.nii.gz
+    mv sub-Iroquoise_ses-01_T1w_reorient_swapped.nii.gz sub-Iroquoise_ses-01_T1w.nii.gz
+
+After the labels are reoriented:
+
+.. image:: ./img/crop/good_orient_mark.png
+    :scale: 50%
+    :align: center
+
+
+********************************************
+How to tilt the orientation of your images ?
+********************************************
+
+Nudging the image with FSLEyes
+##############################
+
+In FSLEyes
+----------
+
+1. Open T1w image
+
+.. code-block:: bash
+
+    fsleyes Orig/sub-Pandore_ses-01_rec-mean_T1w.nii.gz
+
+.. image:: ./img/nudge/Pandore_01_Orig.png
+    :scale: 50%
+    :align: center
+
+
+2. Open Nudge Menu
+
+In the upper bar menu of FSLEyes, click on "Tools" -> "Nudge"
+
+.. image:: ./img/nudge/Pandore_01_NudgeMenu.png
+    :scale: 50%
+    :align: center
+
+
+3. Change to World coordinates
+
+
+Click on "Change display Space"
+
+.. image:: ./img/nudge/Pandore_01_AfterChangeCoordToWorld.png
+    :scale: 50%
+    :align: center
+
+
+4. Rotate the image
+
+Should be: biggest dimension along the anterior-posterior axis (Y)
+No Left/right tilt (equilibrated)
+Facing straight when viewed from from the top
+
+.. image:: ./img/nudge/Pandore_01_AfterRotate.png
+    :scale: 50%
+    :align: center
+
+
+
+5.  Saving the affine
+
+Click on "Save affine" in the lower menu of Nudge window, and
+!!Important!! choose  "FLIRT" for file type and save the affine under your favorite name .mat (e.g nudge_T1w.mat)
+
+.. image:: ./img/nudge/Pandore_01_SaveAffine.png
+    :scale: 50%
+    :align: center
+
+Apply the nudge transfo to the Orig image
+-----------------------------------------
+
+
+.. code-block:: bash
+
[…]
