+++ .github/workflows/check_on_PR.yml
@@ -25,7 +25,7 @@ jobs:
             sudo apt-get install build-essential  graphviz libgraphviz-dev
             pip install --upgrade pygraphviz graphviz
 
-        - name: Test Academic Cloud Connectivity
+      - name: Test Academic Cloud Connectivity
         run: |
             # Test HTTP headers, response code, and redirect behavior
             curl -I -v -L --max-time 15 "https://amubox.univ-amu.fr/public.php/dav/files/qzsfymDYjx692Q4" || true
