+++ macapype/utils/tests/test_utils_nodes.py
@@ -23,6 +23,7 @@ def test_NodeParams_init():
     T2_file = op.join(data_path, "sub-Apache_ses-01_T2w.nii")
 
     params = {"t_file": T1_file, "t2_file": T2_file, "aT2": True}
+
     with pytest.raises(AssertionError):
         bet_crop = NodeParams(interface=T1xT2BET(),
                                 params=params,
