### 2026-08-11 7f36908
Test cloud limitations (#326)
added in github actions check_on_PR.yml a part "Test Academic Cloud Connectivity" before pytest to be sure the dataset used in  unit-tests is available
