+++ macapype/utils/tests/test_utils_nodes.py
@@ -23,10 +23,11 @@ def test_NodeParams_init():
     T2_file = op.join(data_path, "sub-Apache_ses-01_T2w.nii")
 
     params = {"t_file": T1_file, "t2_file": T2_file, "aT2": True}
+
     with pytest.raises(AssertionError):
         bet_crop = NodeParams(interface=T1xT2BET(),
-                                params=params,
-                                name="bet_crop")
+                              params=params,
+                              name="bet_crop")
         bet_crop.run()
 
 
@@ -45,6 +46,7 @@ def test_NodeParams_load_inputs_from_dict():
 
     with pytest.raises(AssertionError):
         bet_crop.load_inputs_from_dict(params)
+
 
 def test_MapNodeParams():
 
