### 2026-08-12 e7a777f
Modif according to adamltyson remarks
