+++ paper_JOSS/paper.bib
@@ -30,8 +30,8 @@ journal={PLoS One}, volume={12}, pages={1--21}, year={2017}, publisher={Public L
     FreeSurfer}, author={Fischl, B}, journal={NeuroImage}, volume={62}, pages={774--781}, year={2012}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2012.01.021}
     }
 
-@book{frackowiak1997human, title={
-    Human Brain Function}, author={Frackowiak, RSJ and Friston, KJ and Frith, CD and Dolan, RJ and Mazziotta, JC}, year={1997}, publisher={Academic Press USA} , url={https://psycnet.apa.org/doi/10.1027/0269-8803.14.2.128}}
+@article{tierney2025SPM, title={
+    SPM 25: open source neuroimaging analysis software}, author={Tierney, Tim and Alexander, N A and others}, journal={Journal of Open Source Software}, year={2025},volume={ 10(110)}, pages={8103}, doi={10.21105/joss.08103}}
 
 @article{garcia2021preemacs, title={
     PREEMACS: Pipeline for preprocessing and extraction of the macaque brain surface}, author={Garcia-Saldivar, P and Garimella, A and others}, fullauthor={Garcia-Saldivar, P and Garimella, A and Garza-Villarreal, EA and Mendez, FA and Concha, L and Merchant, H}, journal={NeuroImage}, volume={227}, pages={117671}, year={2021}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2020.117671}
@@ -56,11 +56,11 @@ year = {2019}
 }
 
 @article{lepage2021civet, title={
-    CIVET-Macaque: An automated pipeline for MRI-based cortical surface generation and cortical thickness in macaques}, author={Lepage, C and Wagstyl, K and others},  fullauthor={Lepage, C and Wagstyl, K and Jung, B and Seidlitz, J and Sponheim, C and Ungerleider, L and Wang, X and Evans, AC and Messinger, A}, journal={Neuroimage}, volume={227}, pages={117622}, year={2021}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2020.117622}
+    CIVET-Macaque: An automated pipeline for MRI-based cortical surface generation and cortical thickness in macaques}, author={Lepage, C and Wagstyl, K and others},  fullauthor={Lepage, C and Wagstyl, K and Jung, B and Seidlitz, J and Sponheim, C and Ungerleider, L and Wang, X and Evans, AC and Messinger, A}, journal={NeuroImage}, volume={227}, pages={117622}, year={2021}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2020.117622}
     }
 
 @article{liu2021marmoset, title={
-    Marmoset Brain Mapping V3: Population multi-modal standard volumetric and surface-based templates}, author={Liu, C and others}, journal={Neuroimage}, year={2021}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2020.117620}
+    Marmoset Brain Mapping V3: Population multi-modal standard volumetric and surface-based templates}, author={Liu, C and others}, journal={NeuroImage}, year={2021}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2020.117620}
     }
 
 @article{lohmeier2019atlasbrex, title={
@@ -71,7 +71,7 @@ year = {2019}
 @article{love2016average, title={The average baboon brain: MRI templates and tissue probability maps from 89 individuals}, author={Love, S and others}, journal={NeuroImage}, volume={132}, year={2016}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2016.03.018} }
 
 @article{messinger2021collaborative, title={
[…]
