+++ paper_JOSS/paper.bib
@@ -114,10 +114,6 @@ year = {2019}
     A coefficient of agreement for nominal scales}, author={Cohen, J}, journal={Educ Psychol Meas}, volume={20}, pages={37--46}, year={1960}, publisher={Sage Publications}, doi={10.1177/001316446002000104}
     }
 
-@phdthesis{martin2002empirical, title={
-    An Empirical Approach to Grouping and Segmentation}, author={Martin, D}, school={University of California, Berkeley, CA, USA}, year={2002} , url={https://dl.acm.org/doi/abs/10.5555/937222}
-    }
-
 @article{Modat2014niftyreg, title={
     Global image registration using a symmetric block- matching approach}, author={Modat,  M and  Cash, DM and Daga, P and Winston, GP and  Duncan, JS and Ourselin, S }, journal={Journal of Medical Imaging}, volume={1(2)}, pages={024003}, year={2014} , doi={10.1117/1.JMI.1.2.024003}
     }
