### 2026-08-13 2dfad9f
sentence case corrections
