+++ paper_JOSS/paper.bib
@@ -128,7 +128,6 @@ year = {2019}
 
 @article{ZHONG2024120652,title = {
     nBEST: Deep-learning-based non-human primates Brain Extraction and Segmentation Toolbox across ages, sites and species},journal = {NeuroImage},volume = {295},pages = {120652},year = {2024},issn = {1053-8119},doi = {10.1016/j.neuroimage.2024.120652},url = {https://www.sciencedirect.com/science/article/pii/S1053811924001472},
-
     author = {Tao Zhong and Xueyang Wu and others},
     fullauthor = {Tao Zhong and Xueyang Wu and Shujun Liang and Zhenyuan Ning and Li Wang and Yuyu Niu and Shihua Yang and Zhuang Kang and Qianjin Feng and Gang Li and Yu Zhang}
     }
