### 2026-08-13 53492e1
NiftyReg and citation nipype
