+++ paper_JOSS/paper.bib
@@ -36,9 +36,10 @@ journal={PLoS One}, volume={12}, pages={1--21}, year={2017}, publisher={Public L
 @article{garcia2021preemacs, title={
     PREEMACS: Pipeline for preprocessing and extraction of the macaque brain surface}, author={Garcia-Saldivar, P and Garimella, A and others}, fullauthor={Garcia-Saldivar, P and Garimella, A and Garza-Villarreal, EA and Mendez, FA and Concha, L and Merchant, H}, journal={NeuroImage}, volume={227}, pages={117671}, year={2021}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2020.117671}
     }
+
 @article{gorgolewski2011nipype,
 title={Nipype: a flexible, lightweight and extensible neuroimaging data processing framework in Python},
-author={Gorgolewski, K and Burns, Christopher and others}, journal={Front. Neuroimform.}, volume={5}, pages={13}, year={2011}, publisher={Frontiers Media SA},  doi={10.3389/fninf.2011.00013} }
+author={Gorgolewski, K and Burns, Christopher and others}, journal={Front. Neuroinform.}, volume={5}, pages={13}, year={2011}, publisher={Frontiers Media SA},  doi={10.3389/fninf.2011.00013} }
 
 @article{gorgolewski2016brain, title={The brain imaging data structure, a format for organizing and describing outputs of neuroimaging experiments}, author={Gorgolewski, K and others}, journal={Sci Data}, volume={3}, pages={160044}, year={2016}, publisher={Nature Publishing Group} , doi={10.1038/sdata.2016.44}}
 
@@ -115,11 +116,13 @@ year = {2019}
     }
 
 @article{Modat2014niftyreg, title={
-    Global image registration using a symmetric block- matching approach}, author={Modat,  M and  Cash, DM and Daga, P and Winston, GP and  Duncan, JS and Ourselin, S }, journal={Journal of Medical Imaging}, volume={1(2)}, pages={024003}, year={2014} , doi={10.1117/1.JMI.1.2.024003}
+    Global image registration using a symmetric block-matching approach}, author={Modat,  M and  Cash, DM and Daga, P and Winston, GP and  Duncan, JS and Ourselin, S }, journal={Journal of Medical Imaging}, volume={1(2)}, pages={024003}, year={2014} , doi={10.1117/1.JMI.1.2.024003}
     }
 
-@article{Wang2021unet, title={
-    U-Net Model for Brain Extraction on Non-human Primates}, author={Wang, X and Li, X and Cho, J W and  Russ, B and  Rajamani, N and Omelchenko, A and Ai, L and  Korchmaros, A and  Garcia, P and Wang, Z and  Kalin, N H and Schroeder, C E and Craddock, C and Fox, A S and  Evans, A and Messinger, A and Milham, M P and & Xu, T }, journal={NeuroImage}, volume={235}, pages={118001}, year={2021} , doi={10.1016/j.neuroimage.2021.118001}
+@article{Wang2021unet, title={U-Net Model for Brain Extraction on Non-human Primates},
+    author={Wang, X and Li, X and others},
+    fullauthor={Wang, X and Li, X and Cho, J W and  Russ, B and  Rajamani, N and Omelchenko, A and Ai, L and  Korchmaros, A and  Garcia, P and Wang, Z and  Kalin, N H and Schroeder, C E and Craddock, C and Fox, A S and  Evans, A and Messinger, A and Milham, M P and Xu, T },
+    journal={NeuroImage}, volume={235}, pages={118001}, year={2021} , doi={10.1016/j.neuroimage.2021.118001}
     }
 
 @article{ZHONG2021117649,title = {
