### 2026-08-17 0fcbce3
Paper joss (#327)
* md, bib and PDF + 2 figures for JOSS submssion

* affiliation

* doi

* licence_MIT

* modif length

* modif length (#315)

* pipelines.jpg

* modified some DOIs

* Modification to take into account new guidelines:Software Design` , `Research Impact Statement` section in paper and `AI usage disclosure` sections

* Related Packages -> State of the field

* Minor spelling and grammatical fixes to JOSS submission (#318)

* Modif according to adamltyson remarks

* suppressed unused ref

* sentence case corrections

* citation

* NiftyReg and citation nipype

* update png file (
[…]
