+++ paper_JOSS/paper.bib
@@ -1,0 +1,136 @@
+@article{avants2011reproducible, title={
+    A reproducible evaluation of ANTs similarity metric performance in brain image registration}, author={Avants, BB and Tustison, NJ and others},
+    fullauthor={Avants, BB and Tustison, NJ and Song, G and Cook, PA and Klein, A and Gee, JC},
+    journal={NeuroImage}, volume={54}, pages={2033--2044}, year={2011}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2010.09.025} }
+
+@article{balbastre2017primatologist, title={
+    Primatologist: A modular segmentation pipeline for macaque brain morphometry}, author={Balbastre, Y and others}, journal={NeuroImage}, volume={162}, pages={306--321}, year={2017}, publisher={Elsevier} , doi={10.1016/j.neuroimage.2017.09.007}}
+
+@article{cieslak2021qsiprep, title={
+    QSIPrep: an integrative platform for preprocessing and reconstructing diffusion MRI data}, author={Cieslak, M and others}, journal={Nature methods}, volume={18}, number={7}, pages={775--778}, year={2021}, publisher={Nature Publishing Group} , doi={10.1038/s41592-021-01185-5}}
+
+@article{esteban2019fmriprep,
+title={fMRIPrep: a robust preprocessing pipeline for functional MRI}, author={Esteban, O and Birman, D and others},
+fullauthor={Esteban, O and Birman, D and Schaer, M and Koyejo, OO and Poldrack, RA and Gorgolewski, KJ},
+journal={Nature methods}, volume={16}, number={1}, pages={111--116}, year={2019}, publisher={Nature Publishing Group} , doi={10.1038/s41592-018-0235-4}
+    }
+
+@article{cox1996afni, title={
+    AFNI: Software for analysis and visualization of functional magnetic resonance neuroimages}, author={Cox, RW}, journal={Computers and Biomedical Research}, volume={29}, pages={162--173}, year={1996}, publisher={Elsevier}, doi={10.1006/cbmr.1996.0014}
+    }
+
+@article{esteban2017mriqc,
+title={ MRIQC: advancing the automatic prediction of image quality in MRI from unseen sites},
+author={Esteban, O and Birman, D and others},
+fullauthor={Esteban, O and Birman, D and Schaer, M and Koyejo, OO and Poldrack, RA and Gorgolewski, KJ},
+journal={PLoS One}, volume={12}, pages={1--21}, year={2017}, publisher={Public Library of Science}, doi={10.1371/journal.pone.0184661}
+    }
+
+@article{fischl2012freesurfer, title={
+    FreeSurfer}, author={Fischl, B}, journal={NeuroImage}, volume={62}, pages={774--781}, year={2012}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2012.01.021}
+    }
+
+@article{tierney2025SPM, title={
+    SPM 25: open source neuroimaging analysis software}, author={Tierney, Tim and Alexander, N A and others}, journal={Journal of Open Source Software}, year={2025},volume={ 10(110)}, pages={8103}, doi={10.21105/joss.08103}}
+
+@article{garcia2021preemacs, title={
+    PREEMACS: Pipeline for preprocessing and extraction of the macaque brain surface}, author={Garcia-Saldivar, P and Garimella, A and others}, fullauthor={Garcia-Saldivar, P and Garimella, A and Garza-Villarreal, EA and Mendez, FA and Concha, L and Merchant, H}, journal={NeuroImage}, volume={227}, pages={117671}, year={2021}, publisher={Elsevier}, doi={10.1016/j.neuroimage.2020.117671}
+    }
+
+@article{gorgolewski2011nipype,
[…]
