+++ _sources/auto_examples/sg_execution_times.rst.txt
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:00.003** total execution time for 1 file **from auto_examples**:
+**00:00.004** total execution time for 1 file **from auto_examples**:
 
 .. container::
 
@@ -33,5 +33,5 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_auto_examples_plot_segment_sphinx_macaque_ants_based.py` (``plot_segment_sphinx_macaque_ants_based.py``)
-     - 00:00.003
+     - 00:00.004
      - 0.0
