+++ _sources/auto_examples/plot_segment_sphinx_macaque_ants_based.rst.txt
@@ -105,7 +105,7 @@ Testing plot in local
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.004 seconds)
+   **Total running time of the script:** (0 minutes 0.003 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_segment_sphinx_macaque_ants_based.py:
