+++ macapype/_version.py
@@ -1,1 +1,1 @@
-__version__ = '0.7-rc2'
+__version__ = '0.7-rc3'
