### 2026-08-17 8880e2b
amubox -> zenodo for macapype ci
