+++ .github/workflows/build_on_prerelease.yml
@@ -86,7 +86,8 @@ jobs:
       - name: Download dataset
         run: |
 
-            curl https://amubox.univ-amu.fr/public.php/dav/files/CoSC5w5ATZextme --output macapype_CI_v2.1.zip
+            curl https://zenodo.org/records/21980768/files/macapype_CI_v2.1.zip?download=1 --output macapype_CI_v2.1.zip
+
 
             unzip -o macapype_CI_v2.1.zip -d macapype_CI_v2.1
 
@@ -130,7 +131,8 @@ jobs:
       - name: Download dataset
         run: |
 
-            curl https://amubox.univ-amu.fr/public.php/dav/files/CoSC5w5ATZextme --output macapype_CI_v2.1.zip
+            curl https://zenodo.org/records/21980768/files/macapype_CI_v2.1.zip?download=1 --output macapype_CI_v2.1.zip
+
 
             unzip -o macapype_CI_v2.1.zip -d macapype_CI_v2.1
 
@@ -174,7 +176,8 @@ jobs:
       - name: Download dataset
         run: |
 
-            curl https://amubox.univ-amu.fr/public.php/dav/files/CoSC5w5ATZextme --output macapype_CI_v2.1.zip
+            curl https://zenodo.org/records/21980768/files/macapype_CI_v2.1.zip?download=1 --output macapype_CI_v2.1.zip
+
 
             unzip -o macapype_CI_v2.1.zip -d macapype_CI_v2.1
 
@@ -218,7 +221,8 @@ jobs:
       - name: Download dataset
         run: |
 
-            curl https://amubox.univ-amu.fr/public.php/dav/files/CoSC5w5ATZextme --output macapype_CI_v2.1.zip
+            curl https://zenodo.org/records/21980768/files/macapype_CI_v2.1.zip?download=1 --output macapype_CI_v2.1.zip
+
 
             unzip -o macapype_CI_v2.1.zip -d macapype_CI_v2.1
 
@@ -263,7 +267,8 @@ jobs:
       - name: Download dataset
         run: |
 
-            curl https://amubox.univ-amu.fr/public.php/dav/files/CoSC5w5ATZextme --output macapype_CI_v2.1.zip
+            curl https://zenodo.org/records/21980768/files/macapype_CI_v2.1.zip?download=1 --output macapype_CI_v2.1.zip
+
 
             unzip -o macapype_CI_v2.1.zip -d macapype_CI_v2.1
 
@@ -308,7 +313,8 @@ jobs:
       - name: Download dataset
         run: |
 
-            curl https://amubox.univ-amu.fr/public.php/dav/files/CoSC5w5ATZextme --output macapype_CI_v2.1.zip
+            curl https://zenodo.org/records/21980768/files/macapype_CI_v2.1.zip?download=1 --output macapype_CI_v2.1.zip
+
             unzip -o macapype_CI_v2.1.zip -d macapype_CI_v2.1
 
 
@@ -352,7 +358,8 @@ jobs:
       - name: Download dataset
         run: |
 
-            curl https://amubox.univ-amu.fr/public.php/dav/files/CoSC5w5ATZextme --output macapype_CI_v2.1.zip
+            curl https://zenodo.org/records/21980768/files/macapype_CI_v2.1.zip?download=1 --output macapype_CI_v2.1.zip
+
 
             unzip -o macapype_CI_v2.1.zip -d macapype_CI_v2.1
 
@@ -396,7 +403,8 @@ jobs:
       - name: Download dataset
         run: |
 
-            curl https://amubox.univ-amu.fr/public.php/dav/files/CoSC5w5ATZextme --output macapype_CI_v2.1.zip
[…]
