+++ .github/workflows/build_on_prerelease.yml
@@ -289,9 +289,9 @@ jobs:
 
             ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/macapype_ants_4animal/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/macapype_ants_4animal_t1t2/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/macapype_ants_4animal/sub-Tresor/ses-01/anat/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/macapype_ants_4animal_t1t2/sub-Tresor/ses-01/anat/
             echo ""
 
 
@@ -380,9 +380,9 @@ jobs:
 
             ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/macapype_ants/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/macapype_ants_t1t2/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/macapype_ants/sub-Prune/ses-3/anat/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/macapype_ants_t1t2/sub-Prune/ses-3/anat/
             echo ""
 
       - name: Cleaning dataset
@@ -390,7 +390,7 @@ jobs:
             sudo rm -rf /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results
 
   run-pretest-ants-baboon3:
-    name: Running all existing pipelines in macapype_CI_v2.1.zip (ANTS_robustreg_prep baboon)
+    name: Running all existing pipelines in macapype_CI_v2.1.zip (ANTS_robustreg_prep baboon3)
     needs:  docker-build
     runs-on: ubuntu-latest
     timeout-minutes: 1440
@@ -425,9 +425,9 @@ jobs:
 
             ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3/derivatives/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3/derivatives/macapype_ants/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3/derivatives/macapype_ants_t1t2/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3/derivatives/macapype_ants/sub-Prune/ses-3/anat/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3/derivatives/macapype_ants_t1t2/sub-Prune/ses-3/anat/
             echo ""
 
       - name: Cleaning dataset
@@ -435,7 +435,7 @@ jobs:
             sudo rm -rf /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3
 
   run-pretest-ants-baboon2:
-    name: Running all existing pipelines in macapype_CI_v2.1.zip (ANTS_robustreg_prep baboon)
+    name: Running all existing pipelines in macapype_CI_v2.1.zip (ANTS_robustreg_prep baboon2)
     needs:  docker-build
[…]
