+++ .github/workflows/build_on_release.yml
@@ -71,7 +71,6 @@ jobs:
           tags: ${{ steps.meta.outputs.tags }}
           labels: ${{ steps.meta.outputs.labels }}
 
-
   run-pretest-ants-macaque:
     name: Running all existing pipelines in macapype_CI_v2.1.zip (ANTS_robustreg_prep macaque)
     needs:  docker-build
@@ -87,6 +86,7 @@ jobs:
         run: |
 
             curl https://zenodo.org/records/21980768/files/macapype_CI_v2.1.zip?download=1 --output macapype_CI_v2.1.zip
+
 
             unzip -o macapype_CI_v2.1.zip -d macapype_CI_v2.1
 
@@ -288,9 +288,9 @@ jobs:
 
             ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/macapype_ants_4animal/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/macapype_ants_4animal_t1t2/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/macapype_ants_4animal/sub-Tresor/ses-01/anat/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_marmo/results/derivatives/macapype_ants_4animal_t1t2/sub-Tresor/ses-01/anat/
             echo ""
 
 
@@ -379,9 +379,9 @@ jobs:
 
             ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/macapype_ants/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/macapype_ants_t1t2/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/macapype_ants/sub-Prune/ses-3/anat/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results/derivatives/macapype_ants_t1t2/sub-Prune/ses-3/anat/
             echo ""
 
       - name: Cleaning dataset
@@ -389,7 +389,7 @@ jobs:
             sudo rm -rf /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results
 
   run-pretest-ants-baboon3:
-    name: Running all existing pipelines in macapype_CI_v2.1.zip (ANTS_robustreg_prep baboon)
+    name: Running all existing pipelines in macapype_CI_v2.1.zip (ANTS_robustreg_prep baboon3)
     needs:  docker-build
     runs-on: ubuntu-latest
     timeout-minutes: 1440
@@ -424,9 +424,9 @@ jobs:
 
             ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3/derivatives/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3/derivatives/macapype_ants/
+            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3/derivatives/macapype_ants_t1t2/
 
-            ls /home/runner/work/macapype/macapype/macapype_CI_v2.1/cerimed_baboon/results3/derivatives/macapype_ants/sub-Prune/ses-3/anat/
[…]
