+++ macapype/_version.py
@@ -1,1 +1,1 @@
-__version__ = '0.7-rc9'
+__version__ = '0.7.1'
