### 2026-08-19 #10203 @davidmeunier79
Hi @adamltyson , I have take into account your latest comments in the manuscript, thanks again for in-depth checking.

I have maid a PR with the paper_JOSS branch containing the sources and the PDF in the main repo, and made a release from after merge of the branch.

macapype version:
0.7.1

DOI of the release on zenodo:
https://doi.org/10.5281/zenodo.21996103

Final checks seems OK for me
