### 2026-08-19 #10203 @adamltyson
Hi @davidmeunier79, could you make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper?
