### 2026-08-20 #10203 @adamltyson
@davidmeunier79 could you also update the archive title to match the paper? I.e. "Macapype: An open multi-software framework for1
non-human primate brain anatomical MRI processing" rather than "macapype"?
