### 2026-08-20 #10203 @adamltyson
Hi @davidmeunier79, apologies I did one last check of the paper and found a spelling mistake in the figure. It should be "skull stripped", not "skull striped". Other than that, all looks good and we can move to acceptance.
