### 2026-08-20 07a7bec
Deploying to gh-pages from @ Macatools/macapype@f487a19f0aaedfe849e14889047c7b6b416f519f 🚀
