### 2026-08-20 #10203 @davidmeunier79
> Hi [@davidmeunier79](https://github.com/davidmeunier79), apologies I did one last check of the paper and found a spelling mistake in the figure. It should be "skull stripped", not "skull striped". Other than that, all looks good and we can move to acceptance.

OK @adamltyson I have modified the Figure accordingly :
https://github.com/Macatools/macapype/blob/master/paper_JOSS/pipelines.png
but I am not sure I can still use the command "generate pdf" (looks like no operation is performed)
