### 2026-08-21 #10203 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
