### 2026-08-30 #10203 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] Gorgolewski reference: ensure {Capitalization} of Python
- [ ] Statement of need: add "packages" after "software"
- [ ] State of the field: either hyphenate both or don't hyphenate both of "deep learning" and "machine learning"
- [ ] same paragraph: "perform" --> "performs"
- [ ] Software design section: remove comma after "different species"
