### 2026-08-31 #10203 @davidmeunier79
Dear @rkurchin , many thanks for your help. I have maid a round of modifications regarding your typo checks, the last generated PDF takes into account your latest remarks
